# src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  title: School API
  version: 1.0.0
  #  description: API para gestionar colegio
  description: |-
    API gestion school, this API is a simple example of how to use OpenAPI 3.0 with Spring Boot.

    Some useful links:
    - [Developer LinkedIn](https://www.linkedin.com)
    - [Repository GitHub](https://github.com/josephmn/school/tree/main/src/main/resources/openapi.yaml)
  termsOfService: https://swagger.io/terms/
  contact:
    email: josephcarlos.jcmn@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
#servers:
#  - url: http://localhost:8082
tags:
  - name: people
    description: Operations about people
paths:
  /people:
    get:
      tags:
        - people
      summary: Get all people.
      description: Get all people.
      operationId: getAllPeople
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponsePeopleDto'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - people
      summary: Create people.
      description: Create people in database.
      operationId: createPeople
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPeopleDto'
        required: true
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePeopleDto'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  #  /people/actives:
  #    get:
  #      tags:
  #        - people
  #      summary: Get all people actives.
  #      description: Get all people status actives.
  #      operationId: getAllPeoplesActives
  #      responses:
  #        '200':
  #          description: successful operation
  #          content:
  #            application/json:
  #              schema:
  #                type: array
  #                items:
  #                  $ref: '#/components/schemas/ResponsePeopleDto'
  #        '400':
  #          description: Invalid username supplied
  #        '404':
  #          description: User not found
  #        default:
  #          description: Unexpected error
  #          content:
  #            application/json:
  #              schema:
  #                $ref: "#/components/schemas/Error"

  /people/{id}:
    get:
      tags:
        - people
      summary: Get people by ID.
      description: Get people by ID.
      operationId: getPeopleById
      parameters:
        - name: id
          in: path
          required: true
          description: Id of people
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePeopleDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - people
      summary: Put people by ID.
      description: Put people by ID.
      operationId: updatePeopleById
      parameters:
        - name: id
          in: path
          required: true
          description: Id of people
          schema:
            type: integer
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPeopleDto'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePeopleDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - people
      summary: Delete people by ID.
      description: Delete people by ID.
      operationId: deletePeopleById
      parameters:
        - name: id
          in: path
          required: true
          description: Id of people
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDTO'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /people/findByName/{name}:
    get:
      tags:
        - people
      summary: Get list people by name.
      description: Get list people by name.
      operationId: getListPeopleByName
      parameters:
        - name: name
          in: path
          required: true
          description: Name of people
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponsePeopleDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /people/findByLastName/{lastName}:
    get:
      tags:
        - people
      summary: Get list people by last name.
      description: Get list people by last name.
      operationId: getListPeopleByLastName
      parameters:
        - name: lastName
          in: path
          required: true
          description: Last name of people
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponsePeopleDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /people/updateByDocument/{document}:
    put:
      tags:
        - people
      summary: Put people by document.
      description: Put people by document.
      operationId: updatePeopleByDocument
      parameters:
        - name: document
          in: path
          required: true
          description: Document of people
          schema:
            type: string
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPeopleDto'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePeopleDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    RequestPeopleDto:
      type: object
      required:
        - type_document
        - number_document
        - name_first
        - name_last
        - date_birth
        - gender
      properties:
        type_document:
          type: string
          pattern: '^(DNI|CE)$'
          example: 'DNI'
          description: Document type (DNI, CE, etc.)
        number_document:
          type: string
          maxLength: 15
          pattern: '^(?!0+$)(?:\d{8}|[a-zA-Z0-9]{9,12})$'
          example: '78653423'
          description: Document number
        name_first:
          type: string
          maxLength: 100
          example: Jhon
          description: First name of people
        name_last:
          type: string
          maxLength: 100
          example: James
          description: Last name of people
        date_birth:
          type: string
          format: date
          example: '1993-12-01'
          description: Birth date of people
        gender:
          type: string
          pattern: '^(M|F)$'
          maxLength: 1
          example: 'M'
          description: Gender (M, F, etc.)
        address:
          type: string
          maxLength: 100
          example: '400 RELLA BOULEVARD'
          description: Address of people
        phone:
          type: string
          maxLength: 15
          pattern: '^\+?[1-9]\d{1,14}$'
          example: '+51998765765'
          description: Phone number of people
        email:
          type: string
          format: email
          maxLength: 50
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
          example: "correo@domain.com"
          description: Email of people
        date_register:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01]) (0[0-9]|1[0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$'
          example: '2025-08-02 07:48:56'
          description: Date when the people was registered

    ResponsePeopleDto:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 1
          description: ID generated for people
        type_document:
          type: string
          pattern: '^(DNI|CE)$'
          example: 'DNI'
          description: Document type (DNI, CE, etc.)
        number_document:
          type: string
          maxLength: 15
          pattern: '^(?!0+$)(?:\d{8}|[a-zA-Z0-9]{9,12})$'
          example: '78653423'
          description: Document number
        name_first:
          type: string
          maxLength: 100
          example: Jhon
          description: First name of people
        name_last:
          type: string
          maxLength: 100
          example: James
          description: Last name of people
        date_birth:
          type: string
          format: date
          example: '1993-12-01'
          description: Birth date of people
        gender:
          type: string
          pattern: '^(M|F)$'
          maxLength: 1
          example: 'M'
          description: Gender (M, F, etc.)
        address:
          type: string
          maxLength: 100
          example: '400 RELLA BOULEVARD'
          description: Address of people
        phone:
          type: string
          maxLength: 15
          pattern: '^\+?[1-9]\d{1,14}$'
          example: '+51998765765'
          description: Phone number of people
        email:
          type: string
          format: email
          maxLength: 50
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
          example: "correo@domain.com"
          description: Email of people
        date_register:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01]) (0[0-9]|1[0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$'
          example: '2025-08-02 07:48:56'
          description: Date when the people was registered

    Error:
      type: object
      properties:
        code:
          type: string
          example: 'XXX'
        message:
          type: string
          example: 'Error message'
      required:
        - code
        - message

    ResponseDTO:
      type: object
      properties:
        code:
          type: string
          example: 'XXX'
        message:
          type: string
          example: 'Message response'
      required:
        - code
        - message